requirements.txt
flask
mysql-connector-python
python-dotenv
flask-cors
pyjwt
werkzeug
